requirements.txt
vllm>=0.2.7
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pyyaml>=6.0
accelerate>=0.24.0
transformers>=4.36.0
psutil>=5.9.0
